Bad casting recipe, no cookie.
When adding the casting recipes for metal tool parts, the liquid metal
quantity is produced from the calculation
`IPattern.getPatternCost(ItemStack) * ingotInMillibuckets / 2`. However,
the ItemStack provided to getPatternCost is supposed to be an ItemStack
containing the pattern (or, in this case, the cast), **not** an
ItemStack containing the tool part.

Also, the name "metalCast" for what is obviously the actual output item
(e.g. pickaxe head) is kind of silly. Let's fix that.