Fix player heads dropping with the wrong metadata
Also, can we get some sort of HeadDropEvent or something like that so I can hook Headcrumbs into this withoug much hassle? <3

I saw the TODO on the top of the class and I must say that a Class->ItemStack registry would not suffice at all. Some entities have more than one head (like the vanilla Skeletons). It would be better if there was an actual fired Event, or some sort of registry for head drop handlers of some sort (a simple class with a ```ItemStack getHead(Entity entity)``` method)