Add seared furnaces, a new multiblock like a smeltery, only needing a frame/ceiling
All the walls and floor be seared blocks. Ceiling may be either that or slabs/stairs. A tank can be placed anywhere in the frame

This commit also fixes an issue with the tooltips for smeltery melting errors and with the multiblock ceiling detection, and moves some stuff related to the smeltery tanks to a common class