Modified the glowing modifier to check for liquid blocks before glowing block placement (#2659)
* Add a entity in water check to the glowing modifer

With this commit, the glowing modifier will no longer trying to add glowing blocks around the entity if the said entity is in water. Fixes #2535

* Modify glowing modifier to place glowing blocks in the night.

* Change canBlockStay to check for BlockLiquid

This check is performed before replacing the block in addition to solid
side check. With this change, the inital change of checking entity in
water is not need anymore.