Fix pickblock on racks and tables losing the wood type and racks missing their NBT facing the wrong way
Racks are rotated by the table model, so the tags in the blockstate are redunadant. Reason they were still there is originally the breaking animation used the data from the blockstate (a.k.a. the state) so it faced the wrong direction.

Since people keep somehow getting it  without NBT and Forge changed it to use the extended state for breaking animations, that work around is no longer needed so I might as well make no NBT racks look fine again.