Add back tool forge recipes
This also prevents the tool forge from creating unneeded oredict entries as recipes have to always be registered. Currently they look a little weird in JEI/the recipe book, but as I was told this is the proper way to add an optional oredict recipe that is on Forge/JEI to fix.