Rewrite crafting station
Instead of running recipe calculations in the container, they are now run in the tile entity. This means we only need to calculate them once instead of once per player.
Results are synced with client players upon the recipe changing or the player opening the container
Also fixes a bug where the result would not disappear if previously visible, and removed the tile entity class name blacklist in favor of registry names