Gradle updates
Update Gradle to 6.8, Forge Gradle to 4.0.8, and Lombok to 5.3.0. Fixes an issue trying to run Gradle in an IDE on newer java versions due to Gradle 5. Downgrading would have required downgrading lombok as well

Seems we still have a problem building on newer java versions, but that might be a FG bug or a sideeffect of targing Java 8.