Fix material registry not syncing to other players on LAN worlds
Why is syncing such a pain?
Just to keep context around, need to skip syncing when on a dedicated server (local channel) as otherwise it overwrites the material registry with all new instances of the exact same data which can cause some unexpected behavior in recipe display and matches. In general its safer to do ID checks but this means if someone does not it won't break.