Implement dynamic modifiers, phase 1
Framework is in place for dynamc modifiers, but most of our modifiers are still static. That will be updated in later commits
Modifiers don't currently support redirects, but that is planned. Have to decide how I want to implement it
Datagenerator for modifiers is not yet written, will be written soon