Adjust bow base stats
On longbows, 3x durability combined with the effective 2x from the heads was a lot. Lowering it to be 1.5x durability and give a bonus 120 so wood ends up at the right value
On crossbows, deciding to say the vanilla crossbow is actually copper and thus do not need to set the base stats. Prevents the need to penalize it in drawspeed when its supposed to be the fast tool