Switch to a much more stable method for fastUseItem
Turns out there is a forge event right next to the spot in vanilla that makes you go slowly while holding a shield. Wonder how I missed it before, I as read that code to figure out how much to make you faster by. Probably caused by having a solution in mind and ignoring anything unrelated to that solution
As part of this, simplify the modifier clearing logic. Will be even more simple if the forge hook gets merged, I should consider implementing a modifier hook for that ahead of time and just call it in the 6 different locations for now