Add tool build modifier hooks
Small semantic change in the removal hooks: before you had onRemove (no return) called early during modifier removal and validate called with 0 late in removal. With the new hooks, onRemove has a return and handles removal validation.
In 1 out of 3 cases, the old hooks will function exactly the same as before. Apparently they were called inconsistently though, so the order of removal vs validations changes slightly for the other two cases.
I do not believe it was possible to rely on the old order hence the breaking change, this new order just makes it so much easier to understand.