Fix indestructable tools still sometimes despawning (#5170)
Technically, it will still potentially despawn if you leave the game running for several years with how this is setup, its the cost to pay for the item spinning. Due to short rounding if the game shuts down then the age of the item will be truncated and thus never reach the lifetime.