Fix too little fire protection on shields and staffs (#5169)
Vanilla only checks the armor slots for protection modifiers so the prescence of it from vanilla was ignored.
Unfortunately, it seems fire tick reduction will not work on shields, not sure if there is anything I can do about that without bringing back the super messy hack or a registry substitution on the enchatment.