Implement damage source predicate and protection modifier modules
Protection module does all your general protection needs, with predicates supported on both the damage source and the target of the damage.
Due to the secondary effects of modifiers not all being migrated, was only able to implement Protection, Feather Falling, Fire Protection, and Turtle Shell
Also included is the Source blocked module, to block damage entirely. Right now its just used for long fall, but in the future its also needed for frost walker

Included in this commit is a utility to make it easier to create a leveling value builder, since that keeps coming up