Implement tag injector for loading book pages on modifiers and tools
This tag loader allows adding new pages to the book by adding them to the tag then putting a page JSON in the appropite place, meaning addons without their own book can add their content to the tinkers books.
Requires moving a lot of JSON files around as names for pages are a lot more important now. Have done so for all books using the new feature.
The one thing not supported is a new armor set outside the encyclopedia as the existing sections did not seem reasonable to extend. Will need to consider whether this is feasibe to do.

As part of this change, migrated The One Probe to a dynamic modifier to make conditioning easier, and it conditionally loads now.
In addition, added pages on enderslime staff to the book