Fix some modiifers and smeltery recipes consuming allowing filled tanks (#5155)
Due to the container implementation, this allowed you to consume fluid instead of the tank itself
Works using a no container ingredient that matches items that lack a container. Primarily used for NBT sensitive containers. Using that for all such cases of dealing with empty containers instead of checking for empty NBT