Translation Updates and Fixes
The translation of "smelt" and "melt" was same due to my insufficient information.
Thus I searched the differences of them and find that "ergit-" is the meaning of "melt".
And "kal" noun is the closest word to "smelt that I could find so I use "kalla-" verb.

The other changes are made to books tooltips.