Compound Word Fixes
TDK says that compund words that has color in it and is related to plant or animal should be contiguous.
Thus, I changed slimes naming and some plants or mushroom naming to be contiguous.

I also make other things contigious but they are not living things.

And other small fixes.