Rework looting hook to have the new features of the enchantment hook
Looting hook now supports predicates to condition over the holder, target, and damage source
Looting hook now has two variants, the main one for weapons (both melee and ranged), and a second one for armor (which works on any slot except the one doing the looting). The default implementations stick to armor slots (meaning chestplates and alike could boost looting if we wished)
TinkerLivingEntityPredicate is now just TinkerPredicate, a class for helpers related to predicates and custom singletons. It also now has some helpers for nullable predicates (null matches any)
After thinking through the logic of this change, it occured to me there may be some conflicts with looting on staffs acting as "armor looting" as staffs are "held armor", so I opted to have our default implementations worn armor only for the enchantments as well