Take advantage of loadables to reduce code for modifier module loaders, variables, and predicates
IdParser is now a loadable for the relevant resource location variant
IntRange is now a loadable for IntRange, it had a weird loading more of itself behavior going already
In addition to those two, added loadables for: SlotType, LevelingValue, RandomLevelingValue, ModifierEntry, SlotType, IToolStat, INumericToolStat, TagKey<Modifier>, MaterialVariantId, plus various enums

As part of this change, variable formulas now store "percent" as a field inside as it made it made it require less context shenanigans for parsing. Since we always want it for the tooltip anyways, does not hurt to place in that interface. They also are far simplier in other ways, notably all the loading logic was moved to a loadable.